made JavaErlang non-static; split each class in its own file
